requirements.txt
ultralytics==8.0.20
Pillow==9.5.0
geopandas==0.13.2
rasterio==1.3.8
numpy==1.24.3
scikit-learn==1.2.2
shapely==2.0.1
pyshp==2.3.1
pyproj==3.6.0
Flask==2.3.2
Werkzeug==2.3.6
